requirements.txt
disutils==1.4.32.post2
Django==5.1.4
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
Jinja2==3.1.6
joblib==1.4.2
matplotlib==3.10.0
mysqlclient==2.2.6
numpy==2.2.1
pandas==2.2.3
pandas-datareader==0.10.0
pillow==11.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.0
setuptools==75.6.0
sqlparse==0.5.3
tf-keras==2.15.0
torch==2.7.1
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
